Added files that manipulate how the week view looks and fixed up the previous week, month, and year view as well
Added a NoCollision file and TimelineGrid file that is just made for the weekview and changed up the css so that adding a course looks much nicer on the week view. As for the other views, the headers were off position and same with the courses, so we changed that through the css as well.